+++ lib/src/git_frame.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/rendering.dart';
+
+class GifFrame {
+  final ImageInfo imageInfo;
+  final Duration duration;
+
+  GifFrame(this.imageInfo, this.duration);
+}